Grepo addition
Hello, I would like to make some changes to the bot but I got stuck on one thing and I probably just need to show to how to do it. If xadam1 or tfojt reads this or anyone who understands the code could you please contact me on discord Orig1nal#3772. Thank you